[tests] Reorder fixtures
- Move Image fixture from tests.instruments.test_models to tests.conftest.
- Order fixtures in tests.conftest alphabetically.
- Refactor tests.conftest remove duplicate code
